[REF] orm: move TransientModel into its own file
Separate the `TransientModel` into its own file to reduce models.py.
