+++ odoo/models/__init__.py
@@ -13,7 +13,6 @@ from odoo.orm.models import (
     BaseModel,
     MetaModel,
     Model,
-    TransientModel,
     check_companies_domain_parent_of,
     check_company_domain_parent_of,
     fix_import_export_id_paths,
@@ -21,6 +20,7 @@ from odoo.orm.models import (
     to_record_ids,
 )
 from odoo.orm.model_classes import is_model_class, is_model_definition
+from odoo.orm.models_transient import TransientModel
 from odoo.orm.table_objects import Constraint, Index, UniqueIndex
 from odoo.orm.utils import (
     READ_GROUP_TIME_GRANULARITY,
